[joystick] new joystick calibration program

+++ sw/ground_segment/joystick/sdl_stick.h
@@ -38,7 +38,7 @@
 #define STICK_INPUT_DEV_MAX   15
 
 /* Global variables about the initialized device */
-extern int8_t stick_axis_values[STICK_AXIS_COUNT];
+extern int16_t stick_axis_values[STICK_AXIS_COUNT];
 extern int32_t stick_button_values;
 extern uint8_t stick_hat_value;
 extern int stick_axis_count, stick_button_count;
